requirements.txt
# Core ML (CPU fallback for CI, GPU used at runtime)
torch>=2.1.0
transformers>=4.37.0
accelerate>=0.26.0

# bitsandbytes — CUDA only, skip in CPU/CI environments
# Install manually on GPU: pip install bitsandbytes>=0.42.0
# bitsandbytes>=0.42.0

# Embeddings
sentence-transformers>=2.3.0

# Vector DB
faiss-cpu>=1.7.4

# Dataset
datasets>=2.16.0

# Evaluation & Utils
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Memory profiling
psutil>=5.9.0
pynvml>=11.5.0

# API Backend
fastapi>=0.100.0
uvicorn>=0.23.0